refactor(drive): file-based versioning for having and ranked mode detection
Both mode_detection.rs files held the version-table dispatcher and the
frozen v0 grammar in one file, unlike the verify_having_range_proof
module next door. Conform them to the file-based versioning scheme:
mode_detection/mod.rs keeps the dispatcher (and ranked_order_key on the
ranked side, which is version-independent and externally referenced),
mode_detection/v0/mod.rs holds the frozen grammar and its private
helpers. The _v0 entry points are re-exported so drive_dispatcher and
the test suites keep their import paths. Pure motion, no behavior
change.
